+++ lib/internal/ffi_headers/ffi_headers.dart
@@ -7,4 +7,5 @@ import "package:system_info/system_info.dart";
 part "src/ffi_h.dart";
 part "src/ffi_platforms.dart";
 part "src/ffitarget_h.dart";
+part "src/headers.dart";
 part "src/utils.dart";